move assertions to util_test.go
