+++ x/gamm/pool-models/balancer/util_test.go
@@ -1,6 +1,7 @@
 package balancer_test
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
@@ -11,6 +12,7 @@ import (
 	tmtypes "github.com/tendermint/tendermint/proto/tendermint/types"
 	dbm "github.com/tendermint/tm-db"
 
+	"github.com/osmosis-labs/osmosis/v7/osmoutils"
 	"github.com/osmosis-labs/osmosis/v7/x/gamm/pool-models/balancer"
 	"github.com/osmosis-labs/osmosis/v7/x/gamm/types"
 )
@@ -35,4 +37,52 @@ func createTestContext(t *testing.T) sdk.Context {
 	ms := rootmulti.NewStore(db, logger)
 
 	return sdk.NewContext(ms, tmtypes.Header{}, false, logger)
+}
+
+func assertExpectedSharesErrRatio(t *testing.T, expectedShares, actualShares sdk.Int) {
+	allowedErrRatioDec, err := sdk.NewDecFromStr(allowedErrRatio)
+	require.NoError(t, err)
+
+	errTolerance := osmoutils.ErrTolerance{
+		MultiplicativeTolerance: allowedErrRatioDec,
+	}
+
+	require.Equal(
+		t,
+		0,
+		errTolerance.Compare(expectedShares, actualShares),
+		fmt.Sprintf("expectedShares: %s, actualShares: %s", expectedShares.String(), actualShares.String()))
+}
+
+func assertExpectedLiquidity(t *testing.T, tokensJoined, liquidity sdk.Coins) {
+	require.Equal(t, tokensJoined, liquidity)
+}
+
+// assertPoolStateNotModified asserts that sut (system under test) does not modify
+// pool state.
+func assertPoolStateNotModified(t *testing.T, pool *balancer.Pool, sut func()) {
+	// We need to make sure that this method does not mutate state.
+	oldPoolAssets := pool.GetAllPoolAssets()
+	oldLiquidity := pool.GetTotalPoolLiquidity(sdk.Context{})
+	oldShares := pool.GetTotalShares()
+
+	sut()
+
+	newPoolAssets := pool.GetAllPoolAssets()
+	newLiquidity := pool.GetTotalPoolLiquidity(sdk.Context{})
+	newShares := pool.GetTotalShares()
+
+	require.Equal(t, oldPoolAssets, newPoolAssets)
+	require.Equal(t, oldLiquidity, newLiquidity)
+	require.Equal(t, oldShares, newShares)
+}
+
+// assertPanic if expectPanic is true, asserts that sut (system under test)
+// panics. If expectPanic is false, asserts that sut does not panic.
+func assertPanic(t *testing.T, expectPanic bool, sut func()) {
+	if expectPanic {
+		require.Panics(t, sut)
+	} else {
+		require.NotPanics(t, sut)
+	}
 }
